refactor: Background 스타일 분리
