+++ src/components/background/Background.tsx
@@ -1,20 +1,6 @@
 /** @jsxImportSource @emotion/react */
-import { css } from '@emotion/react';
 import BackgroundImg from './BackgroundImg';
-
-const backgroundStyle = css({
-    width: '100vw',
-    height: '100vh',
-    display: 'flex',
-    position: 'fixed',
-    background: 'black',
-    img: {
-        objectFit: 'cover',
-        height: '100vh',
-        width: '100vw',
-        opacity: 0.8,
-    },
-});
+import { backgroundStyle } from './Style';
 
 const Background = () => {
     return (
